Improvements for the SPEC MockTurtle node template
- Port 'sfp_mod_def0' was named with '_b' but defined as input. I changed the name and I added a
default value, so that cards without an SFP do no need to drive any value.

- I have regrouped all the pieces of code withing a "if( g_with_white_rabbit ) generate" statement.
So that a user can firgure out faster what it is and it is not instantiated if this generic is false.

- I have moved "cmp_clk_vcxo : BUFG" at line 655, inside a "if( g_with_white_rabbit ) generate"
statement together with the "cmp_dmtd_clk_pll : PLL_BASE" and the "cmp_clk_dmtd_buf : BUFG" since
I believe that they only need ot be instantiated if g_with_white_rabbit=TRUE. Please check!

- Default input value added to clk_20m_vcxo_i port (@ line 100) so that the designs that do not
use it do not have to drive this signal.

- Default input value added to fmc0_clk_aux_i port (@line 191) so that the designs that do not
use it do not have to drive this signal.

- Buffer 'U_Buf_CLK_GTP' moved to line 839 since the signal 'clk_125m_gtp' is only used by the
'wr_gtp_phy_spartan6' module which is not instantiated if
(g_with_wr_phy and g_with_white_rabbit)=FALSE. Now is together to the 'wr_gtp_phy_spartan6'
instance that is the only place that uses this gtp signal.

- As consequence, default values have been added to clk_125m_gtp_n_i/p__i, so that the designs
that do not use it do not have to drive this signal.Lines 105 and 106

- Added a generic parameter for the value to load in the DAC controlling the 25MHz VCXO in case
that the WR core is not instantiated (so DACs not disciplined by it). Lines 87-89.

- Added a "U_DAC_ARB : spec_serial_dac_arb" instance inside a "if ( not g_with_white_rabbit )
generate" block to drive the SPEC DAC signals in case the WR core is not instantiated. It uses
the value passed as integer through the generic g_fixed_dac25MHz. Added also a mono-stable
process for the 'load' signal. Lines 938-976.

-----------------------------------------------------------------------------------------
   spec_node_pkg.vhd
-----------------------------------------------------------------------------------------

- Port 'sfp_mod_def0_b' renamed as 'sfp_mod_def0_i' since it is declared as input.
A default value (:='0') was added so that cards without an SFP do no need to drive any value.

- Added value by default (:='0') to 'clk_20m_vcxo_i' port.

- Added value by default (:='1') to 'clk_125m_gtp_n_i' port.

- Added value by default (:='0') to 'clk_125m_gtp_p_i' port.

- Added a generic parameter: "g_fixed_dac25MHz  : integer := 39321" for programming the
25MHz VCXO DAC if g_with_white_rabbit=false. It sets up the VCXO control voltage to its middle level.